learn/pyproject.toml
[project]
name = "cybersecurity-incident-predictor"
version = "0.1.0"
description = "Advanced ensemble ML platform for predicting cybersecurity incidents"
authors = [{name = "CyberSec ML Team", email = "team@cybersec-ml.com"}]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    # Core ML and Data Processing
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "torch-geometric>=2.4.0",
    "xgboost>=2.0.0",
    "lightgbm>=4.1.0",
    "scikit-learn>=1.4.0",
    "transformers>=4.35.0",
    "polars>=0.19.0",
    "pyarrow>=14.0.0",
    "pydantic>=2.5.0",
    "optuna>=3.4.0",
    "numpy>=1.24.0",
    # Dashboard and Visualization
    "streamlit",
    "plotly>=5.17.0",
    "anywidget>=0.8.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    # Model Interpretability
    "umap-learn>=0.5.0",
    "scikit-image>=0.22.0",
    # Data Handling
    "pandas>=2.1.0",
    "datasets>=2.14.0",
    "kaggle>=1.5.16",
    # Development and Quality
    "pytest>=7.4.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
    "black>=23.9.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
    "typer>=0.9.0",
    "rich>=13.6.0",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
    "notebook>=7.0.0",
]

deployment = [
    "docker>=6.1.0",
    "redis>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.scripts]
cybersec-dashboard = "src.dashboard.app:main"
cybersec-train = "src.pipeline.training_pipeline:main"
cybersec-predict = "src.pipeline.inference_pipeline:main"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*" = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["src"]
force-sort-within-sections = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
